requirements.txt
fastapi===0.134.0
uvicorn === 0.41.0
requests === 2.32.3
python-dotenv ===  1.0.1
jinja2 === 3.1.4